requirements.txt
Django
django-import_export
djangorestframework